app/src/main/java/com/example/huff6/localdatabasecalllog/DatabaseHelper.java
package com.example.huff6.localdatabasecalllog;

import android.content.ContentValues;
import android.content.Context;
import android.database.Cursor;
import android.database.sqlite.SQLiteDatabase;
import android.database.sqlite.SQLiteOpenHelper;

/**
 *
 */
public class DatabaseHelper extends SQLiteOpenHelper {
    public static final String DATABASE_NAME = "callLog.db";
    public static final String TABLE_NAME = "callLog_table";
    //public static final String COL_1 = "ID";
    public static final String COL_2 = "NAME";
    public static final String COL_3 = "PHONE";
    public static final String COL_4 = "START_TIME";
    public static final String COL_5 = "END_TIME";
    public static final String COL_6 = "DURATION";
    public static final String COL_7 = "NOTES";

    public DatabaseHelper(Context context) {
        super(context, DATABASE_NAME, null, 1);
    }

    @Override
    public void onCreate(SQLiteDatabase db) {
        db.execSQL("create table " + TABLE_NAME + " (ID INTEGER PRIMARY KEY AUTOINCREMENT," +
                "NAME TEXT,PHONE TEXT,START_TIME TEXT,END_TIME TEXT,DURATION TEXT,NOTES TEXT)");
        //db.execSQL("CREATE TABLE log_table (id INTEGER PRIMARY KEY, name TEXT, phone TEXT);");
    }

    @Override
    public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
        db.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
        onCreate(db);
    }

    public boolean insertData(String name,    String number,   String startTime,
                              String endTime, String duration, String notes) {
        SQLiteDatabase db = this.getWritableDatabase();
        ContentValues contentValues = new ContentValues();
        contentValues.put(COL_2, name);
        contentValues.put(COL_3, number);
        contentValues.put(COL_4, startTime);
        contentValues.put(COL_5, endTime);
        contentValues.put(COL_6, duration);
        contentValues.put(COL_7, notes);

        long result = db.insert(TABLE_NAME, null, contentValues);
        return result != -1;
    }

    public Cursor getAllData() {
        SQLiteDatabase db = this.getWritableDatabase();
        return db.rawQuery("select * from " + TABLE_NAME, null);
    }

    /*
    public Integer deleteData(String id) {
        SQLiteDatabase db = this.getWritableDatabase();
        return db.delete(TABLE_NAME, "ID = ?", new String[] { id });
    }
    */

    public void deleteAllData() {
        SQLiteDatabase db = this.getWritableDatabase();
        //db.execSQL("delete from " + TABLE_NAME);
        db.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
        onCreate(db);
    }
}
